♻️ Refactor filter
- make useFilter hook
