+++ src/pages/problem/long/LongProblemListPage.tsx
@@ -5,11 +5,10 @@ import PageTemplate from '../../../templates/PageTemplate';
 import { URL } from '../../../constants/url';
 import { ProblemTable } from '../../../components/Table/ProblemTable';
 import { ILongProblem } from '../../../types/problem/api';
-import { HeadCell, IFilter, IProblemCondition } from '../../../types/etc';
+import { HeadCell } from '../../../types/etc';
 import { longProblemApiWrapper } from '../../../api/wrapper/problem/longProblemApiWrapper';
-import { useState, MouseEvent, useEffect, ChangeEvent } from 'react';
-import { FILTER_PARAMS, PROBLEM_FILTER } from '../../../constants/filter';
-import { v4 as uuidv4 } from 'uuid';
+import { PROBLEM_FILTER } from '../../../constants/filter';
+import { useFilter } from '../../../hooks/useFilter';
 
 const headCells: readonly HeadCell[] = [
   {
@@ -67,38 +66,10 @@ const tableHeads: (keyof ILongProblem)[] = [
 ];
 
 export const LongProblemListPage = () => {
-  const [filterState, setFilterState] = useState<IFilter[]>([]);
+  const { filterState, addFilter, deletetFilter, updateCondition, updateFilterValue } = useFilter();
 
   function getLongDataList(page: number, params: object) {
     return longProblemApiWrapper.getLongProblemList({ ...params, page: page });
-  }
-
-  function addFilter() {
-    setFilterState((prev) => [...prev, { id: uuidv4(), condition: 'id', value: '' }]);
-  }
-  function deletetFilter(event: MouseEvent<Element, MouseEvent>) {
-    const id = event.currentTarget.id;
-    setFilterState((prev) => prev.filter((e) => e.id !== id));
-  }
-
-  function updateCondition(newCondition: IProblemCondition, DOMId: string) {
-    setFilterState((prev) =>
-      prev.map(({ id, condition, value }) =>
-        id === DOMId
-          ? ({ id, value, condition: newCondition } as unknown as IFilter)
-          : { id, condition, value },
-      ),
-    );
-  }
-  function updateFilterValue(event: ChangeEvent<HTMLTextAreaElement>) {
-    if (!event.currentTarget) return;
-    const DOMId = event.currentTarget.id;
-    const DOMValue = event.currentTarget.value;
-    setFilterState((prev) =>
-      prev.map(({ id, condition, value }) =>
-        id === DOMId ? { id, value: DOMValue, condition } : { id, condition, value },
-      ),
-    );
   }
 
   return (
